requirements.txt
crispy-bootstrap5==0.7
Django==4.2.3
daphne==3.0.2
django-bootstrap-icons==0.8.3
django-bootstrap-v5==1.0.11
django-crispy-forms==2.0
djangorestframework==3.14.0
Pillow==10.0.0